+++ src/components/markdown_editor/index.ts
@@ -17,23 +17,19 @@
  * under the License.
  */
 
-import React, { ComponentType } from 'react';
-
-export const EuiIcon = ({ type, ...rest }: any) => (
-  <div
-    data-euiicon-type={
-      typeof type === 'string' ? type : type.displayName || type.name
-    }
-    {...rest}
-  />
-);
-
-export const appendIconComponentCache = (_: {
-  [iconType: string]: ComponentType;
-}) => {
-  // manually appending to the internal EuiIcon cache is out-of-scope of this test environment
-};
-
-export const TYPES = [];
-export const COLORS = [];
-export const SIZES = [];
+export { EuiMarkdownEditor, EuiMarkdownEditorProps } from './markdown_editor';
+export {
+  getDefaultEuiMarkdownParsingPlugins,
+  getDefaultEuiMarkdownProcessingPlugins,
+} from './plugins/markdown_default_plugins';
+export { EuiMarkdownContext } from './markdown_context';
+export { EuiMarkdownFormat } from './markdown_format';
+export {
+  EuiMarkdownParseError,
+  EuiMarkdownAstNode,
+  EuiMarkdownAstNodePosition,
+  EuiMarkdownFormatting,
+  EuiMarkdownEditorUiPlugin,
+  RemarkRehypeHandler,
+  RemarkTokenizer,
+} from './markdown_types';