+++ src/webpack.config.js
@@ -22,7 +22,7 @@
 const path = require('path');
 const webpack = require('webpack');
 const CircularDependencyPlugin = require('circular-dependency-plugin');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+const TerserPlugin = require('terser-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
 const isProduction = process.env.NODE_ENV === 'production';
@@ -45,6 +45,10 @@ const plugins = [
     maxChunks: 1,
   }),
 ];
+
+const terserPlugin = new TerserPlugin({
+  sourceMap: true,
+});
 
 module.exports = {
   mode: isProduction ? 'production' : 'development',
@@ -99,16 +103,7 @@ module.exports = {
   },
 
   plugins,
+  optimization: {
+    minimizer: isProduction ? [terserPlugin] : [],
+  },
 };
-
-if (isProduction) {
-  const optimization = (module.exports.optimization =
-    module.exports.optimization || {});
-  optimization.minimizer = [
-    new UglifyJsPlugin({
-      uglifyOptions: {
-        sourceMap: true,
-      },
-    }),
-  ];
-}